# openapi-mod-demo/openapi/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Generated API",
    "version" : "1.0"
  },
  "paths" : {
...
    "/resteasy-jackson/test/{parameter}" : {
      "get" : {
        "tags" : [ "Test Resource" ],
        "operationId" : "TestResource_list",
        "parameters" : [ {
          "name" : "parameter",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-MyHeader",
          "in" : "header",
          "description" : "*bla*",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "demo" : [ "Role1", "Role2" ]
        }, {
          "demo" : [ "Role1", "Role2" ]
        } ]
      },
      "post" : {
        "tags" : [ "Test Resource" ],
        "operationId" : "TestResource_list2",
        "parameters" : [ {
          "name" : "parameter",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-MyHeader",
          "in" : "header",
          "description" : "*bla*",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Quark"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Quark" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "demo" : {
        "type" : "http",
        "description" : "You login",
        "scheme" : "basic"
      }
    }
  }
}